+++ src/Application/Command/PromotionCode/DecrementStock.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Proximum\Vimeet\Application\Command\PromotionCode;
+
+use Proximum\Vimeet\Domain\Model\PromotionCode;
+
+class DecrementStock
+{
+    /** @var PromotionCode */
+    public $promotionCode;
+
+    public function __construct(PromotionCode $promotionCode)
+    {
+        $this->promotionCode = $promotionCode;
+    }
+}